+++ packages/vue-elementui-boilerplate/src/components/system/form/cascader/index.vue
@@ -26,7 +26,7 @@
 
 <script>
 export default {
-  name: 'NTCascader',
+  name: 'VRNCascader',
   components: {},
   props: {
     data: {
