+++ packages/vue-elementui-boilerplate/src/components/system/form/index.js
@@ -1,12 +1,12 @@
 /*
  * @Author: Cphayim
  * @Date: 2019-05-19 22:27:16
- * @LastEditTime: 2019-06-06 10:21:06
+ * @LastEditTime: 2020-07-07 09:02:09
  * @Description: 组件、结构化数据类聚合出口
  */
 
 // 导出 NTForm 组件
-export { default as NTForm } from './nt-form.vue'
+export { default as VRNForm } from './vrn-form.vue'
 
 // 导出所有表单控件组件
 export * from './controls'
