+++ packages/vue-elementui-docker-boilerplate/src/components/system/form/checkbox/index.vue
@@ -48,7 +48,7 @@
 
 <script>
 export default {
-  name: 'NTCheckbox',
+  name: 'VRNCheckbox',
   components: {},
   props: {
     data: {
