+++ packages/vue-elementui-docker-boilerplate/src/components/system/form/radio/index.vue
@@ -48,7 +48,7 @@
 
 <script>
 export default {
-  name: 'NTRadio',
+  name: 'VRNRadio',
   components: {},
   props: {
     data: {
